Fix matching_fee to double

+++ lib/screens/male/screens/buy_service/components/buy_service_cancel_confirmation_dialog.dart
@@ -7,7 +7,7 @@ class BuyServiceCancelConfirmationDialog extends StatelessWidget {
     this.matchingFee,
   }) : super(key: key);
 
-  final int matchingFee;
+  final double matchingFee;
 
   @override
   Widget build(BuildContext context) {
